Render the shield plate's base geometry, no folds/tabs/handle/patterns yet
First Shield rendering slice, mirroring the banner flag's own first
geometry-only slice: draws the shield plate's six-face cuboid from the
version's default shield base texture, using pr-35-head's source
dimensions (12x22x1). Deliberately excludes the handle, pattern
stamping, interaction regions, and fold/tab guides.

Extracted the texture-frame/rounding/BaseMinecraft pipeline banner.ts
already had into a new shapes/shared.ts, since the shield plate needs
the same draw-from-shared-atlas logic; banner.ts now consumes it
instead of duplicating it. banner.test.ts's generic rounding-function
tests moved to shared.test.ts alongside their new home.

Verified: types/lint/check:imports clean, vitest 13/13 (unchanged,
moved), this generator's own Playwright spec 7/7 (generator-content
change, not the full suite).
