+++ src/generators/minecraftBannerAndShield/minecraftBannerAndShieldGenerator.tsx
@@ -32,6 +32,7 @@ import {
   drawBannerPattern,
   drawBannerPole,
 } from "./shapes/banner";
+import { drawShieldPlate } from "./shapes/shield";
 import titleImage from "./images/title-a4.png";
 
 const id = "minecraft-banner-and-shield";
@@ -41,7 +42,8 @@ const name = "Minecraft Banner and Shield";
 const instructions: InstructionsDef = `
 Component-by-component rebuild in progress. The banner flag base, pole, and
 crossbar are rendered with fold and tab guides, and clicking the flag
-arms/stamps the selected pattern; shield follows in a separate slice.
+arms/stamps the selected pattern. The shield plate renders its base geometry
+only so far, with no fold/tab guides, handle, or pattern stamping yet.
 `;
 
 const images: ImageDef[] = [{ id: "Title", url: titleImage.src }];
@@ -83,6 +85,8 @@ const render = (ctx: RenderContext, props: BannerAndShieldProps): void => {
       props.showFolds
     );
     ctx.defineRegion(bannerFlagFrontRegion(), bannerFlagRegionId);
+  } else {
+    drawShieldPlate(ctx, props.versionId);
   }
 
   // Temporary: proves the Title overlay is wired end to end. Removed once
